+++ packages/life_hooks/lib/life_hooks.dart
@@ -1,3 +1,5 @@
 library life_hooks;
 
+export 'src/abstract/abstract.dart';
+export 'src/hooks/hooks.dart';
 export 'src/life_hook.dart';