+++ packages/life_hooks/lib/src/abstract/primitives.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/widgets.dart';
+
+abstract class Loadable {
+  Loadable();
+
+  /// Use this function to load something on
+  /// instance initialization
+  Future<void> onLoad({required final BuildContext context});
+}